deploy and un deploy functionality (#3073)
* deploy and undeploy functionality


* shown response in the UI


---------

+++ src/serveressFunction/build-run-deploy.ts
@@ -4,13 +4,16 @@
  *-----------------------------------------------------------------------------------------------*/
 
 import { EventEmitter, Terminal, Uri, commands, window } from 'vscode';
-import { ClusterVersion, FunctionObject } from './types';
+import { ClusterVersion, FunctionContent, FunctionObject, RunResponse } from './types';
+import validator from 'validator';
 import { ServerlessCommand, Utils } from './commands';
 import { Platform } from '../util/platform';
 import { OdoImpl } from '../odo';
 import { CliChannel } from '../cli';
 import { ChildProcess, SpawnOptions } from 'child_process';
 import { ServerlessFunctionView } from './view';
+import { multiStep } from './multiStepInput';
+import { Progress } from '../util/progress';
 
 export class BuildAndDeploy {
 
@@ -46,22 +49,22 @@ export class BuildAndDeploy {
         }
     }
 
-    public async buildFunction(functionName: string, functionUri: Uri): Promise<void> {
-        const exisitingTerminal = this.buildTerminalMap.get(`build-${functionUri.fsPath}`);
-        const outputEmitter = this.buildEmiterMap.get(`build-${functionUri.fsPath}`);
+    public async buildFunction(context: FunctionObject): Promise<void> {
+        const exisitingTerminal = this.buildTerminalMap.get(`build-${context.folderURI.fsPath}`);
+        const outputEmitter = this.buildEmiterMap.get(`build-${context.folderURI.fsPath}`);
         if (exisitingTerminal) {
             let exisitingProcess = this.buildPrcessMap.get(exisitingTerminal);
-            void window.showWarningMessage(`Do you want to restart ${functionName} build ?`, 'Yes', 'No').then(async (value: string) => {
+            void window.showWarningMessage(`Do you want to restart ${context.name} build ?`, 'Yes', 'No').then(async (value: string) => {
                 if (value === 'Yes') {
                     exisitingTerminal.show(true);
                     await commands.executeCommand('workbench.action.terminal.clear')
-                    outputEmitter.fire(`Start Building ${functionName} \r\n`);
+                    outputEmitter.fire(`Start Building ${context.name} \r\n`);
                     exisitingProcess.kill('SIGINT')
                     this.buildPrcessMap.delete(exisitingTerminal);
                     const clusterVersion: ClusterVersion | null = await this.checkOpenShiftCluster();
-                    const buildImage = await this.getImage(functionUri);
-                    const opt: SpawnOptions = { cwd: functionUri.fsPath };
-                    void CliChannel.getInstance().spawnTool(ServerlessCommand.buildFunction(functionUri.fsPath, buildImage, clusterVersion), opt).then((cp) => {
+                    const buildImage = await this.getImage(context.folderURI);
+                    const opt: SpawnOptions = { cwd: context.folderURI.fsPath };
+                    void CliChannel.getInstance().spawnTool(ServerlessCommand.buildFunction(context.folderURI.fsPath, buildImage, clusterVersion), opt).then((cp) => {
                         exisitingProcess = cp;
                         this.buildPrcessMap.set(exisitingTerminal, cp);
                         exisitingProcess.on('error', (err) => {
@@ -75,29 +78,31 @@ export class BuildAndDeploy {
                             void window.showErrorMessage(`${errChunk as string}`);
                         });
                         exisitingProcess.on('exit', () => {
+                            context.hadBuilt = true;
+                            ServerlessFunctionView.getInstance().refresh(context);
                             outputEmitter.fire('\r\nPress any key to close this terminal\r\n');
                         });
                     });
                 }
             });
         } else {
-            await this.buildProcess(functionUri, functionName);
+            await this.buildProcess(context);
         }
     }
 
-    private async buildProcess(functionUri: Uri, functionName: string) {
+    private async buildProcess(context: FunctionObject) {
         const clusterVersion: ClusterVersion | null = await this.checkOpenShiftCluster();
-        const buildImage = await this.getImage(functionUri);
+        const buildImage = await this.getImage(context.folderURI);
         const outputEmitter = new EventEmitter<string>();
         let devProcess: ChildProcess;
         let terminal = window.createTerminal({
-            name: `Build ${functionName}`,
+            name: `Build ${context.name}`,
             pty: {
                 onDidWrite: outputEmitter.event,
                 open: () => {
-                    outputEmitter.fire(`Start Building ${functionName} \r\n`);
-                    const opt: SpawnOptions = { cwd: functionUri.fsPath };
-                    void CliChannel.getInstance().spawnTool(ServerlessCommand.buildFunction(functionUri.fsPath, buildImage, clusterVersion), opt).then((cp) => {
+                    outputEmitter.fire(`Start Building ${context.name} \r\n`);
+                    const opt: SpawnOptions = { cwd: context.folderURI.fsPath };
+                    void CliChannel.getInstance().spawnTool(ServerlessCommand.buildFunction(context.folderURI.fsPath, buildImage, clusterVersion), opt).then((cp) => {
                         this.buildPrcessMap.set(terminal, cp);
                         devProcess = cp;
                         devProcess.on('spawn', () => {
@@ -113,6 +118,8 @@ export class BuildAndDeploy {
                             outputEmitter.fire(`\x1b[31m${errChunk as string}\x1b[0m`.replaceAll('\n', '\r\n'));
                         });
                         devProcess.on('exit', () => {
+                            context.hadBuilt = true;
+                            ServerlessFunctionView.getInstance().refresh(context);
                             outputEmitter.fire('\r\nPress any key to close this terminal\r\n');
                         });
                     });
@@ -121,8 +128,8 @@ export class BuildAndDeploy {
                     if (devProcess && devProcess.exitCode === null) { // if process is still running and user closed terminal
                         devProcess.kill('SIGINT');
                     }
-                    this.buildTerminalMap.delete(`build-${functionUri.fsPath}`);
-                    this.buildEmiterMap.delete(`build-${functionUri.fsPath}`);
+                    this.buildTerminalMap.delete(`build-${context.folderURI.fsPath}`);
+                    this.buildEmiterMap.delete(`build-${context.folderURI.fsPath}`);
                     this.buildPrcessMap.delete(terminal);
                     terminal = undefined;
                 },
@@ -137,8 +144,8 @@ export class BuildAndDeploy {
                 })
             },
         });
-        this.buildTerminalMap.set(`build-${functionUri.fsPath}`, terminal);
-        this.buildEmiterMap.set(`build-${functionUri.fsPath}`, outputEmitter);
+        this.buildTerminalMap.set(`build-${context.folderURI.fsPath}`, terminal);
+        this.buildEmiterMap.set(`build-${context.folderURI.fsPath}`, outputEmitter);
     }
 
     public runFunction(context: FunctionObject, runBuild = false) {
@@ -161,7 +168,15 @@ export class BuildAndDeploy {
                         });
                         runProcess.stdout.on('data', (chunk) => {
                             outputEmitter.fire(`${chunk as string}`.replaceAll('\n', '\r\n'));
-                            void commands.executeCommand('openshift.Serverless.refresh', context);
+                            try {
+                                const json = JSON.parse(`${chunk as string}`) as RunResponse;
+                                if (json.msg?.indexOf('Server listening at') !== -1) {
+                                    outputEmitter.fire(`\n\n Press ctrl+C for stop running ${context.name}`);
+                                }
+                            } catch (err) {
+                                // no action
+                            }
+                            void commands.executeCommand('openshift.Serverless.refresh');
                         });
                         runProcess.stderr.on('data', (errChunk) => {
                             outputEmitter.fire(`\x1b[31m${errChunk as string}\x1b[0m`.replaceAll('\n', '\r\n'));
@@ -177,15 +192,17 @@ export class BuildAndDeploy {
                     }
                     this.runTerminalMap.delete(`run-${context.folderURI.fsPath}`);
                     terminal = undefined;
-                    void commands.executeCommand('openshift.Serverless.refresh', context);
+                    void commands.executeCommand('openshift.Serverless.refresh');
                 },
-                handleInput: ((_data: string) => {
-                    if (!runProcess) {
-                        terminal.dispose();
-                    } else {
-                        outputEmitter.fire('^C\r\n');
-                        runProcess.kill('SIGINT');
-                        terminal.dispose()
+                handleInput: ((data: string) => {
+                    if (data.charCodeAt(0) === 3 || data.charCodeAt(0) === 94) {
+                        if (!runProcess) {
+                            terminal.dispose();
+                        } else {
+                            outputEmitter.fire('^C\r\n');
+                            runProcess.kill('SIGINT');
+                            terminal.dispose()
+                        }
                     }
                 })
             },
@@ -200,23 +217,150 @@ export class BuildAndDeploy {
         }
     }
 
+    public undeployFunction(context: FunctionObject) {
+        void Progress.execFunctionWithProgress(`Undeploying the function ${context.name}`, async () => {
+            const result = await OdoImpl.Instance.execute(ServerlessCommand.undeployFunction(context.name));
+            if (result.error) {
+                void window.showErrorMessage(result.error.message);
+            } else {
+                void commands.executeCommand('openshift.Serverless.refresh');
+            }
+        });
+    }
+
     public async deployFunction(context: FunctionObject) {
         const currentNamespace: string = ServerlessFunctionView.getInstance().getCurrentNameSpace();
         const yamlContent = await Utils.getFuncYamlContent(context.folderURI.fsPath);
         if (yamlContent) {
-            const depyedNamespace = yamlContent.deploy?.namespace;
-            if (depyedNamespace && depyedNamespace !== currentNamespace) {
-                const response = await window.showInformationMessage(`Function namespace (declared in func.yaml) is different from the current active namespace. Deploy function ${context.name} to namespace ${depyedNamespace}?`,
+            const deployedNamespace = yamlContent.deploy?.namespace || undefined;
+            if (!deployedNamespace) {
+                await this.deployProcess(context, deployedNamespace, yamlContent);
+            } else if (deployedNamespace !== currentNamespace) {
+                const response = await window.showInformationMessage(`Function namespace (declared in func.yaml) is different from the current active namespace. Deploy function ${context.name} to namespace ${deployedNamespace}?`,
                     'Ok',
                     'Cancel');
-                if(response === 'Cancel') {
-                    return;
+                if (response === 'Ok') {
+                    await this.deployProcess(context, deployedNamespace, yamlContent);
                 }
-                if (!yamlContent.image || !BuildAndDeploy.imageRegex.test(yamlContent.image)) {
-                    void window.showErrorMessage(`Function ${context.name} has invalid imaage`)
-                }
+            } else {
+                await this.deployProcess(context, deployedNamespace, yamlContent);
             }
         }
+    }
+
+    private async deployProcess(context: FunctionObject, deployedNamespace: string, yamlContent: FunctionContent) {
+        if (!yamlContent.image || !BuildAndDeploy.imageRegex.test(yamlContent.image)) {
+            void window.showErrorMessage(`Function ${context.name} has invalid imaage`)
+            return;
+        }
+        const clusterVersion: ClusterVersion | null = await this.checkOpenShiftCluster();
+        const buildImage = await this.getImage(context.folderURI);
+        const outputEmitter = new EventEmitter<string>();
+        let deployProcess: ChildProcess;
+        let terminal = window.createTerminal({
+            name: `Deploying the function ${context.name}`,
+            pty: {
+                onDidWrite: outputEmitter.event,
+                open: () => {
+                    outputEmitter.fire(`Deploying ${context.name} \r\n`);
+                    const opt: SpawnOptions = { cwd: context.folderURI.fsPath };
+                    void CliChannel.getInstance().spawnTool(ServerlessCommand.deployFunction(context.folderURI.fsPath, buildImage, deployedNamespace, clusterVersion), opt).then((cp) => {
+                        deployProcess = cp;
+                        deployProcess.on('spawn', () => {
+                            terminal.show();
+                        });
+                        deployProcess.on('error', (err) => {
+                            void window.showErrorMessage(err.message);
+                        });
+                        // eslint-disable-next-line @typescript-eslint/no-misused-promises
+                        deployProcess.stdout.on('data', async (chunk) => {
+                            const response = `${chunk as string}`.replaceAll('\n', '\r\n');
+                            if (response.includes('Please provide credentials for image registry')) {
+                                response.replace(/Please provide credentials for image registry/gm, '');
+                                await this.provideUserNameAndPassword(deployProcess, 'Please provide credentials for image registry.');
+                            }
+                            if (response.includes('Incorrect credentials, please try again')) {
+                                response.replace(/Incorrect credentials, please try again/gm, '');
+                                await this.provideUserNameAndPassword(deployProcess, 'Please provide credentials for image registry.', true);
+                            }
+                            outputEmitter.fire(response);
+                        });
+                        deployProcess.stderr.on('data', (errChunk) => {
+                            outputEmitter.fire(`\x1b[31m${errChunk as string}\x1b[0m`.replaceAll('\n', '\r\n'));
+                        });
+
+                        deployProcess.on('exit', () => {
+                            void commands.executeCommand('openshift.Serverless.refresh');
+                            outputEmitter.fire('\r\nPress any key to close this terminal\r\n');
+                        });
+                    });
+                },
+                close: () => {
+                    if (deployProcess && deployProcess.exitCode === null) { // if process is still running and user closed terminal
+                        deployProcess.kill('SIGINT');
+                    }
+                    terminal = undefined;
+                    void commands.executeCommand('openshift.Serverless.refresh');
+                },
+                handleInput: ((_data: string) => {
+                    if (!deployProcess) {
+                        terminal.dispose();
+                    } else {
+                        outputEmitter.fire('^C\r\n');
+                        deployProcess.kill('SIGINT');
+                        terminal.dispose()
+                    }
+                })
+            },
+        });
+    }
+
+    private async provideUserNameAndPassword(
+        process: ChildProcess,
+        message: string,
+        reattemptForLogin?: boolean,
+    ): Promise<void> {
+        const userInfo = 'Provide username for image registry.';
+        const userName = await this.getUsernameOrPassword(
+            message,
+            userInfo,
+            false,
+            'Provide an username for image registry.',
+            reattemptForLogin,
+        );
+        if (!userName) {
+            process.stdin.end();
+            return null;
+        }
+        const passMessage = 'Provide password for image registry.';
+        const userPassword = await this.getUsernameOrPassword(message, passMessage, true, 'Provide a password for image registry.');
+        if (!userPassword) {
+            process.stdin.end();
+            return null;
+        }
+        process.stdin.write(`${userName}\n${userPassword}\n`);
+    }
+
+    private async getUsernameOrPassword(
+        message: string,
+        infoMessage: string,
+        passwordType?: boolean,
+        errorMessage?: string,
+        reattemptForLogin?: boolean,
+    ): Promise<string | null> {
+        return multiStep.showInputBox({
+            title: message,
+            prompt: infoMessage,
+            placeholder: infoMessage,
+            reattemptForLogin,
+            validate: (value: string) => {
+                if (validator.isEmpty(value)) {
+                    return errorMessage;
+                }
+                return null;
+            },
+            password: passwordType,
+        });
     }
 
     public getDefaultImages(name: string): string[] {
